refactor(pack): refactor ecofax activation
