+++ packages/manager/apps/telecom/src/app/telecom/pack/slots/voipEcoFax/activation/activation.router.js
@@ -1,0 +1,11 @@
+import controller from './activation.controller';
+import template from './activation.html';
+
+export default /* @ngInject */ ($stateProvider) => {
+  $stateProvider.state('telecom.packs.pack.voipEcoFax-activation', {
+    url: '/voipEcoFax/activation',
+    template,
+    controller,
+    controllerAs: 'PackFaxActivationCtrl',
+  });
+};
